Allow skipping writeme update with --no-update
